requirements.txt
python-dotenv==1.0.1
openai==0.27.10
